--- styles.js
@@ -1,9 +1,0 @@
-'use strict';
-
-module.exports = {
-  facebook_login_button: {
-    alignItems: 'center',
-    justifyContent: 'center',
-		flex: 1,
-  }
-};
